TS: SearchByIDA
